+++ src/orchestrators/group-orchestrator/orchestrator/src/group_orchestrator/agents/agent_gateway.py
@@ -1,8 +1,8 @@
-import requests
-from typing import Any, Dict
+from typing import Any
 
-from pydantic import BaseModel, ConfigDict
+import aiohttp
 from opentelemetry.propagate import inject
+from pydantic import BaseModel
 
 
 class AgentGateway(BaseModel):
@@ -14,8 +14,12 @@ class AgentGateway(BaseModel):
         protocol = "https" if self.secure else "http"
         return f"{protocol}://{self.host}/{agent_name}/{agent_version}"
 
-    def invoke_agent(
-        self, agent_name: str, agent_version: str, agent_input: BaseModel
+    async def invoke_agent(
+        self,
+        session: aiohttp.ClientSession,
+        agent_name: str,
+        agent_version: str,
+        agent_input: BaseModel,
     ) -> Any:
         payload = agent_input.model_dump_json()
 
@@ -24,13 +28,10 @@ class AgentGateway(BaseModel):
             "Content-Type": "application/json",
         }
         inject(headers)
-        response = requests.post(
+        async with session.post(
             self._get_endpoint_for_agent(agent_name, agent_version),
-            headers=headers,
             data=payload,
-        )
-        response_json = response.json()
-        if response_json:
-            return response_json
-        else:
-            raise Exception("Unable to invoke agent")
+            headers=headers,
+        ) as response:
+            response.raise_for_status()
+            return await response.json()
