+++ src/orchestrators/group-orchestrator/orchestrator/src/group_orchestrator/plan_manager.py
@@ -1,6 +1,7 @@
 from typing import List
 
-from group_orchestrator.agents import PlanningAgent, BaseAgent
+from group_orchestrator.agents import BaseAgent
+from group_orchestrator.agents.planning_agent import PlanningAgent
 from group_orchestrator.go_types import Plan
 
 
@@ -9,7 +10,9 @@ class PlanManager:
         self.planning_agent = planning_agent
 
     async def generate_plan(
-        self, overall_goal: str, task_agents: List[BaseAgent]
+        self,
+        overall_goal: str,
+        task_agents: List[BaseAgent],
     ) -> Plan:
         gen_plan_response = await self.planning_agent.generate_plan(
             overall_goal, task_agents
